% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fay_wu.R
\name{a1}
\alias{a1}
\title{Calculate a sub n, a component of the variance}
\usage{
a1(n)
}
\arguments{
\item{n}{an integer value for the number of chromosomes}
}
\value{
a single numeric value for the estimate of a1
}
\description{
Calculate a sub n, a component of the variance
}
