% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfs.R
\name{aggregate_sfs}
\alias{aggregate_sfs}
\title{Aggreage an n-dimensional SFS into a single dimension}
\usage{
aggregate_sfs(sfs, ...)
}
\arguments{
\item{sfs}{an SFS}

\item{...}{ignored}
}
\value{
an one-dimensional SFS
}
\description{
Aggreage an n-dimensional SFS into a single dimension
}

