% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfs.R
\name{coarsen_sfs}
\alias{coarsen_sfs}
\title{Summarize SFS into 5 bins on each axis}
\usage{
coarsen_sfs(sfs, ...)
}
\arguments{
\item{sfs}{an SFS}

\item{...}{ignored}
}
\value{
an SFS of reduced dimension, which summarizes the full SFS into bins corresponding to singletons, doubletons, all other polymorhisms
	(of higher sample frequency) and fixed differences.
}
\description{
Summarize SFS into 5 bins on each axis
}
\details{
The "coarse-grained" SFS produced by this function is useful as a lower-dimensinoal summary statistic in ABC procedures.
}
\references{
Tellier A, et al. (2011) Estimating parameters of speciation models based on refined summaries of the joint site-frequency spectrum.
		PLoS One 6: e18155.
}

