% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{derivedCount}
\alias{derivedCount}
\title{Get the count of derived (ALT) alleles from a row in a vcf}
\usage{
derivedCount(row)
}
\arguments{
\item{row}{A vector of data in VCF format corresponding to 1 SNP}
}
\value{
a single integer of the number of derived alleles in a
	sample at this SNP
}
\description{
addition to orig sfsr pkg by L.Cooper 11-2017
}
