% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfs.R
\name{theta_w}
\alias{d_xy}
\alias{theta_pi}
\alias{theta_w}
\alias{theta_zeta}
\title{Calculate Watterson's estimator of theta}
\usage{
theta_w(sfs, persite = FALSE, ...)

theta_pi(sfs, persite = FALSE, ...)

theta_zeta(sfs, persite = FALSE, ...)

d_xy(sfs, persite = FALSE, ...)
}
\arguments{
\item{sfs}{an SFS}

\item{persite}{logical; if \code{TRUE}, divide by number of sites}

\item{...}{ignored}
}
\value{
Watterson's \eqn{\theta}, the population-scaled mutation rate

value of \eqn{\hat{\pi} = \sum_{i>j}^N{d_{ij}}}, the average number of pairwise differences (aka expected heterozygosity)

value of \eqn{\hat{\theta_{\zeta}}}, the number of derived singletons in a sample

average pairwise divergence: if one population, to the outgroup (assuming unfolded, polarized SFS);
if two populations, between them
}
\description{
Calculate Watterson's estimator of theta

Calculate the Tajima's pi-hat estimator of theta (avg pairwise differences)

Calculate the Fu and Li's estimator of theta

Calculate the d_xy, the average pairwise divergence
}
\references{
Tajima F (1983) Evolutionary relationship of DNA sequences in finite populations. Genetics 105: 437-360.

Fu YX and Li WH (1993) Statistical tests of neutrality of mutations. Genetics 133: 693-709.
}

