% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfs.R
\name{f_st}
\alias{f_st}
\title{Calculate F_st (genetic differentiation among populations)}
\usage{
f_st(x, bootstraps = TRUE, ...)
}
\arguments{
\item{x}{an SFS with >1 dimension}

\item{bootstraps}{logical; if \code{TRUE} (the default), run calculation on any available bootstraps}

\item{...}{ignored}
}
\value{
estimate of \eqn{F_{st}}, via Weir and Cockerham's classic estimator \eqn{\hat{\theta}}
}
\description{
Calculate F_st (genetic differentiation among populations)
}
\references{
Weir BS, Cockerham C (1984) Estimating F-statistics for the analysis of population structure. Evolution 38: 1358-1370.
}

