% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fay_wu.R
\name{fayWu_H}
\alias{fayWu_H}
\title{Calculate Normalize Fay and Wu's H}
\usage{
fayWu_H(sfs)
}
\arguments{
\item{sfs}{an SFS}
}
\value{
a single numeric value for the estimate of H
}
\description{
Calculate Normalize Fay and Wu's H
}
