% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{flatten_sfs}
\alias{flatten_sfs}
\title{Convert SFS to vector representation for saving in a file}
\usage{
flatten_sfs(sfs)
}
\arguments{
\item{sfs}{an SFS}
}
\value{
a vector containing the 'flattened' representation of the SFS, in row-major order
}
\description{
Convert SFS to vector representation for saving in a file
}

