% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfs.R
\name{hka_test}
\alias{hka_test}
\title{Perform the HKA test.}
\usage{
hka_test(x, y = NULL, simulate = FALSE, nsims = 1000, alpha = 0.05,
  scale = 1, ...)
}
\arguments{
\item{x}{an SFS, or a list thereof (one per locus)}

\item{y}{another SFS; if \code{x} not a list, this must be provided}

\item{simulate}{logical; if \code{TRUE}, get p-values by coalescent simulations instead of analytical approximation}

\item{nsims}{number of simulations to perform (if \code{simulate == TRUE})}

\item{alpha}{target significance level for parameter estimates (if \code{simulate == TRUE})}

\item{scale}{the scaling factor for \eqn{\theta}; 1 = autosome, 3/4 = X chromosome, 1/4 = Y chromosome or mitochondria}

\item{...}{ignored}
}
\value{
a list of class \code{"htest"} with test statistic, degrees of freedom, p-value, and estimates of coalescent parameters
}
\description{
Perform the HKA test.
}
\references{
Hudson RR, Kreitman M, Aguadé M (1987) A test of neutral molecular evolution based on nucleotide data. Genetics 116: 153-159.
}

