% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfs.R
\name{marginalize_sfs}
\alias{marginalize_sfs}
\title{Get an SFS with reduced dimensions}
\usage{
marginalize_sfs(sfs, dims = 1, persite = FALSE, ...)
}
\arguments{
\item{sfs}{an SFS}

\item{dims}{indices over which to sum up allele frequencies}

\item{persite}{logical; if \code{TRUE}, divide by the total number of sites}

\item{...}{ignored}
}
\value{
an SFS of reduced dimension, produced by "marginalizing out" (ie. just summing over) one or more populations
}
\description{
Get an SFS with reduced dimensions
}

