% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfs.R
\name{mask_corners}
\alias{mask_corners}
\title{Zero out frequency classes of fixed for ancestral or derived allele, leaving only polymorphic classes}
\usage{
mask_corners(sfs, ...)
}
\arguments{
\item{sfs}{an SFS}

\item{...}{ignored}
}
\value{
the original SFS with fixed bins (frequency zero or N) set to zero
}
\description{
Zero out frequency classes of fixed for ancestral or derived allele, leaving only polymorphic classes
}

