% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfs.R
\name{tajimaD}
\alias{achazY}
\alias{fuliD}
\alias{fuliF}
\alias{tajimaD}
\title{Calculate Tajima's D}
\usage{
tajimaD(x, ...)

fuliD(x, ...)

fuliF(x, ...)

achazY(x, ...)
}
\arguments{
\item{x}{an SFS}

\item{...}{ignored}
}
\description{
Calculate Tajima's D

Calculate Fu and Li's D (with outgroup)

Calculate Fu and Li's F (with outgroup)

Calculate Achaz's Y* statistic
}
\references{
Tajima F (1989) Statistical method for testing the neutral mutation hypothesis by DNA polymorphism. Genetics 105: 437-460.

Fu YX and Li WH (1993) Statistical tests of neutrality of mutations. Genetics 133: 693-709.

Fu YX and Li WH (1993) Statistical tests of neutrality of mutations. Genetics 133: 693-709.

Achaz G (2008) Testing for neutrality in samples with sequencing errors. Genetics 179: 1409-1424
}

