% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfs.R
\name{noboots}
\alias{noboots}
\title{Discard bootstrap replicates from an SFS}
\usage{
noboots(sfs)
}
\arguments{
\item{sfs}{an SFS}
}
\value{
an SFS of same dimensions with any bootstrap replicates discarded
}
\description{
Discard bootstrap replicates from an SFS
}

