% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read.vcf}
\alias{read.vcf}
\title{Read in VCF format file into R as a table}
\usage{
read.vcf(file, special.char = "##", ...)
}
\arguments{
\item{file}{a file in uncompressed VCF format}
}
\value{
a Data Frame with VCF data
}
\description{
addition to orig sfsr pkg by L.Cooper 11-2017
}
