% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{sfs.fromVcf}
\alias{sfs.fromVcf}
\title{Calculate either a 1 or 2 dimensional SFS from VCF format data}
\usage{
sfs.fromVcf(vcf.list)
}
\arguments{
\item{vcf.list}{a list of length>=1 & length<3 VCF format data frames
use read.vcf to read VCF file into a table/data frame before using this fxn}
}
\value{
either a named Vector (1D) or Matrix (2D) that is an SFS
}
\description{
addition to orig sfsr pkg by L.Cooper 11-2017
}
