% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fay_wu.R
\name{variance.H}
\alias{variance.H}
\title{Calculate the variance of H}
\usage{
variance.H(n, s, theta)
}
\arguments{
\item{n}{an integer value for the number of chromosomes}

\item{s}{an integer value for the number of segregating sites}

\item{theta}{an estimate of theta (best to use thetaW)}
}
\value{
a single number value for the variance
}
\description{
Calculate the variance of H
}
